Bugfixes
auto build now ignores if the only edits are in img and in README

Fixed the Shield up animation on spoofed shield items being cancelled.

Fixed a typo in AbiltiesHud.java

Fixed a typo in the lootcounter actionbar detection
