+++ .github/workflows/build.yml
@@ -6,11 +6,15 @@ on:
       - "*"
     paths-ignore:
       - ".gitignore"
+      - "img"
+      - "README.md"
   pull_request:
     branches:
       - "*"
     paths-ignore:
       - ".gitignore"
+      - "img"
+      - "README.md"
   workflow_dispatch:
 jobs:
   build:
