+++ src/main/java/ch/njol/unofficialmonumentamod/features/strike/ChestCountOverlay.java
@@ -59,7 +59,7 @@ public class ChestCountOverlay extends HudElement {
 
 	public void onActionbarReceived(Text text) {
 		//first one is non-edited the second one is for edited by vlado's counter mod.
-		if (text.getString().equals("+1 Chest added to lootroom.") || text.getString().matches("\u00a76+1 Chest \u00a7cadded to lootroom\\..*")) {
+		if (text.getString().equals("+1 Chest added to lootroom.") || text.getString().matches("\u00a76\\+1 Chest \u00a7cadded to lootroom\\..*")) {
 			currentCount++;
 		}
 	}
