+++ src/main/java/ch/njol/unofficialmonumentamod/mixins/ElytraFeatureRendererMixin.java
@@ -1,6 +1,7 @@
 package ch.njol.unofficialmonumentamod.mixins;
 
 import ch.njol.unofficialmonumentamod.UnofficialMonumentaModClient;
+import ch.njol.unofficialmonumentamod.features.spoof.TextureSpoofer;
 import net.minecraft.client.render.entity.feature.ElytraFeatureRenderer;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ElytraItem;
@@ -14,10 +15,11 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 public class ElytraFeatureRendererMixin<T extends LivingEntity> {
 	@ModifyVariable(method = "render(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/entity/LivingEntity;FFFFFF)V", at = @At(value = "STORE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
 	private ItemStack editStack(ItemStack value) {
-		if (UnofficialMonumentaModClient.spoofer.spoofedItems.containsKey(value.getName().getString().toLowerCase()) &&
-			    !(Registry.ITEM.get(UnofficialMonumentaModClient.spoofer.spoofedItems.get(value.getName().getString().toLowerCase()).getItemIdentifier()) instanceof ElytraItem)) {
+		if (UnofficialMonumentaModClient.spoofer.spoofedItems.containsKey(TextureSpoofer.getKeyOf(value)) &&
+			    !(Registry.ITEM.get(UnofficialMonumentaModClient.spoofer.spoofedItems.get(TextureSpoofer.getKeyOf(value)).getItemIdentifier()) instanceof ElytraItem)) {
 			return value;
 		}
-		return UnofficialMonumentaModClient.spoofer.apply(value);
+		ItemStack edited = UnofficialMonumentaModClient.spoofer.apply(value);
+		return edited != null ? edited : value;
 	}
 }
